Refactored component
