+++ src/component/context.rs
@@ -5,23 +5,14 @@ use std::{
 };
 
 use crate::{
+    component::EventHandler,
     engine::{Command, CommandExecutor},
     render::DrawContext,
     state::{use_effect, HookDependency},
-    View, ViewWrapper,
+    View,
 };
 
-pub type Component = Arc<dyn ComponentImpl>;
-pub type EventHandler = Arc<Mutex<dyn FnMut(&Arc<Context>, &dyn Any) + Send + Sync>>;
-
-pub trait ComponentImpl: Send + Sync {
-    fn call(&self, cx: &Arc<Context>) -> View;
-}
-
-pub struct Scope {
-    pub children: Mutex<Vec<(Arc<Context>, Option<ViewWrapper>)>>,
-    executor: Arc<dyn CommandExecutor>,
-}
+use super::{scope::Scope, Component, ComponentImpl};
 
 pub struct Context {
     component: Mutex<Component>,
@@ -187,49 +178,5 @@ impl Context {
 
     pub fn stop(self: &Arc<Self>) -> crate::Result<()> {
         self.execute(crate::engine::commands::Stop)
-    }
-}
-
-impl Scope {
-    pub fn new(executor: Arc<dyn CommandExecutor>) -> Arc<Self> {
-        Arc::new(Self {
-            children: Mutex::new(Vec::new()),
-            executor,
-        })
-    }
-
-    pub fn child<F: ComponentImpl + 'static>(
-        self: &Arc<Self>,
-        child: F,
-        view_wrapper: Option<ViewWrapper>,
-    ) {
-        let ctx = Context::new(child, self.executor.clone());
-
-        ctx.refresh();
-
-        self.children.lock().unwrap().push((ctx, view_wrapper));
-    }
-
-    pub fn view(self: &Arc<Self>, view: View) {
-        let ctx = Context::new(view, self.executor.clone());
-
-        ctx.refresh();
-
-        self.children.lock().unwrap().push((ctx, None));
-    }
-}
-
-impl ComponentImpl for View {
-    fn call(&self, _: &Arc<Context>) -> View {
-        self.clone()
-    }
-}
-
-impl<F> ComponentImpl for F
-where
-    F: Fn(&Arc<Context>) -> View + Send + Sync,
-{
-    fn call(&self, ctx: &Arc<Context>) -> View {
-        self(ctx)
     }
 }
